fix: update Google login to use GET method
