+++ nuxt-app/components/LoginSignUpDialog.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-  import { emailSignUp, nuxtEmailLogin, nuxtGoogleLogin } from "@/apis/auth";
+  import { emailSignUp, nuxtEmailLogin } from "@/apis/auth";
   import { ElMessage } from "element-plus";
   import { useAuthStore } from "@/stores/auth";
   import { Message, Lock, User } from "@element-plus/icons-vue";
@@ -76,26 +76,6 @@
       email: form.value.email,
       password: form.value.password
     });
-    if (error.value) {
-      ElMessage({
-        message: error.value?.data.message,
-        type: "error"
-      });
-      return;
-    }
-    if (data.value?.token) {
-      await refreshSession();
-      ElMessage({
-        message: `歡迎 ${user.value?.name}`,
-        type: "success"
-      });
-      authStore.setToken(data.value?.token);
-      emit("update:visible", false);
-    }
-  };
-
-  const handleGoogleLogin = async () => {
-    const { data, error } = await nuxtGoogleLogin();
     if (error.value) {
       ElMessage({
         message: error.value?.data.message,
@@ -226,15 +206,19 @@
         </el-button>
         <el-button
           size="large"
-          class="w-full flex text-md text-neutral-800 bg-white hover:border-neutral-800 m-0"
-          @click="handleGoogleLogin"
+          class="w-full text-md text-neutral-800 bg-white hover:border-neutral-800 m-0"
         >
-          <img
-            :src="GoogleSVG"
-            alt="Google Logo"
-            class="w-5 h-5 mr-5 rounded-full"
-          />
-          使用 Google 繼續
+          <a
+            href="https://dev-api.spirvita.tw/api/v1/auth/google"
+            class="flex items-center border px-14 py-2"
+          >
+            <img
+              :src="GoogleSVG"
+              alt="Google Logo"
+              class="w-5 h-5 mr-5 rounded-full"
+            />
+            使用 Google 繼續
+          </a>
         </el-button>
         <p
           class="text-sm cursor-pointer underline"
